+++ ChangeLog.rst
@@ -4,6 +4,7 @@
 0.3.9
 -----
 - [add] Question 和 Answer 添加 deleted 属性
+- [fix] 修复了问题没有回答时 Question.answers 出错的问题
 
 0.3.8
 -----